Added support for the virtio-scsi interface type for storage devices,
thanks to a patch by Boris Jakubith <runkharr@googlemail.com> .

Futhermore the spelling of Nativ was fixed and changed to Native
throughout the source code. Small technical changes were made to the
patch before commiting it.

The patch included the following comment:
This patch adds a new item in the 'Interface' combobox in the "New
Storage Device" dialog: 'virtio-scsi'. In order to make the use of the
Virtio SCSI interface, i had to modify a lot in 'VM.cpp', as the virtual
SCSI device must occur in the Qemu argument list before any drive which
uses it. This interface allows for using "thin provisioning" (aka
"Discard") together with the speed of a para-virtualized storage
interface, so i HAD to add it.